Change variable type

+++ src/hdRunner/main.cpp
@@ -553,7 +553,7 @@ int main(int argc, const char* argv[])
                                                                                    // stratified sampling
     ctx->mSettingsManager->setAs<uint32_t>("render/pt/tonemapperType", 0); // 0 - reinhard, 1 - aces, 2 - filmic
     ctx->mSettingsManager->setAs<uint32_t>("render/pt/debug", 0); // 0 - none, 1 - normals
-    ctx->mSettingsManager->setAs<uint32_t>("render/cameraSpeed", 1);
+    ctx->mSettingsManager->setAs<float>("render/cameraSpeed", 1);
     ctx->mSettingsManager->setAs<float>("render/pt/upscaleFactor", 0.5f);
     ctx->mSettingsManager->setAs<bool>("render/pt/enableUpscale", true);
     ctx->mSettingsManager->setAs<bool>("render/pt/enableAcc", true);
